+++ vue/components/ui/atoms/button-color/button-color.stories.js
@@ -30,7 +30,8 @@ storiesOf("Components/Atoms/Button Color", module)
                         Blue: "blue",
                         Yellow: "yellow",
                         Orange: "orange",
-                        White: "white"
+                        White: "white",
+                        Transparent: "transparent"
                     },
                     "default"
                 )